+++ Runtime/Behaviours/CinemachineBrain.cs
@@ -26,6 +26,49 @@ using UnityEngine.Serialization;
 namespace Cinemachine
 {
     /// <summary>
+    /// This interface is specifically for Timeline.  Do not use it.
+    /// </summary>
+    public interface ICameraOverrideStack
+    {
+        /// <summary>
+        /// Override the current camera and current blend.  This setting will trump
+        /// any in-game logic that sets virtual camera priorities and Enabled states.
+        /// This is the main API for the timeline.
+        /// </summary>
+        /// <param name="overrideId">Id to represent a specific client.  An internal
+        /// stack is maintained, with the most recent non-empty override taking precenence.
+        /// This id must be > 0.  If you pass -1, a new id will be created, and returned.
+        /// Use that id for subsequent calls.  Don't forget to
+        /// call ReleaseCameraOverride after all overriding is finished, to
+        /// free the OverideStack resources.</param>
+        /// <param name="camA">The camera to set, corresponding to weight=0.</param>
+        /// <param name="camB">The camera to set, corresponding to weight=1.</param>
+        /// <param name="weightB">The blend weight.  0=camA, 1=camB.</param>
+        /// <param name="deltaTime">Override for deltaTime.  Should be Time.FixedDelta for
+        /// time-based calculations to be included, -1 otherwise.</param>
+        /// <returns>The override ID.  Don't forget to call ReleaseCameraOverride
+        /// after all overriding is finished, to free the OverideStack resources.</returns>
+        int SetCameraOverride(
+            int overrideId,
+            ICinemachineCamera camA, ICinemachineCamera camB,
+            float weightB, float deltaTime);
+
+        /// <summary>
+        /// See SetCameraOverride.  Call ReleaseCameraOverride after all overriding
+        /// is finished, to free the OverrideStack resources.
+        /// </summary>
+        /// <param name="overrideId">The ID to released.  This is the value that
+        /// was returned by SetCameraOverride</param>
+        void ReleaseCameraOverride(int overrideId);
+
+        /// <summary>
+        /// Get the current definition of Up.  May be different from Vector3.up.
+        /// </summary>
+        Vector3 DefaultWorldUp { get; }
+    }
+
+
+    /// <summary>
     /// CinemachineBrain is the link between the Unity Camera and the Cinemachine Virtual
     /// Cameras in the scene.  It monitors the priority stack to choose the current
     /// Virtual Camera, and blend with another if necessary.  Finally and most importantly,
@@ -45,7 +88,7 @@ namespace Cinemachine
     [AddComponentMenu("Cinemachine/CinemachineBrain")]
     [SaveDuringPlay]
     [HelpURL(Documentation.BaseURL + "manual/CinemachineBrainProperties.html")]
-    public class CinemachineBrain : MonoBehaviour
+    public class CinemachineBrain : MonoBehaviour, ICameraOverrideStack
     {
         /// <summary>
         /// When enabled, the current camera and blend will be indicated in the 
@@ -146,15 +189,27 @@ namespace Cinemachine
             get
             {
                 if (m_OutputCamera == null && !Application.isPlaying)
-#if UNITY_2019_2_OR_NEWER
-                    TryGetComponent(out m_OutputCamera);
-#else
-                    m_OutputCamera = GetComponent<Camera>();
-#endif
+                    ControlledObject.TryGetComponent(out m_OutputCamera);
                 return m_OutputCamera;
             }
         }
         private Camera m_OutputCamera = null; // never use directly - use accessor
+        
+        /// <summary>
+        /// CinemachineBrain controls this GameObject.  Normally, this is the GameObject to which 
+        /// the CinemachineBrain component is attached.  However, it is possible to override this
+        /// by setting this property to another GameObject.  If a Camera component is attached to the 
+        /// Controlled Object, then that Camera component's lens settings will also be driven 
+        /// by the CinemachineBrain.
+        /// If this property is set to null, then CinemachineBrain is controlling the GameObject 
+        /// to which it is attached.  The value of this property will always report as non-null.
+        /// </summary>
+        public GameObject ControlledObject
+        {
+            get => m_TargetOverride == null ? gameObject : m_TargetOverride;
+            set => m_TargetOverride = value;
+        }
+        private GameObject m_TargetOverride = null; // never use directly - use accessor
 
         /// <summary>Event with a CinemachineBrain parameter</summary>
         [Serializable] public class BrainEvent : UnityEvent<CinemachineBrain> {}
@@ -211,7 +266,6 @@ namespace Cinemachine
             if (mFrameStack.Count == 0)
                 mFrameStack.Add(new BrainFrame());
 
-            m_OutputCamera = GetComponent<Camera>();
             CinemachineCore.Instance.AddActiveBrain(this);
             CinemachineDebug.OnGUIHandlers -= OnGuiHandler;
             CinemachineDebug.OnGUIHandlers += OnGuiHandler;
@@ -249,6 +303,7 @@ namespace Cinemachine
         {
             m_LastFrameUpdated = -1;
             UpdateVirtualCameras(CinemachineCore.UpdateFilter.Late, -1f);
+            ControlledObject.TryGetComponent(out m_OutputCamera);
         }
 
         private void OnGuiHandler()
@@ -637,8 +692,9 @@ namespace Cinemachine
                 // No active virtal camera.  We create a state representing its position
                 // and call the callback, but we don't actively set the transform or lens
                 var state = CameraState.Default;
-                state.RawPosition = transform.position;
-                state.RawOrientation = transform.rotation;
+                var target = ControlledObject.transform;
+                state.RawPosition = target.position;
+                state.RawOrientation = target.rotation;
                 state.Lens = LensSettings.FromCamera(m_OutputCamera);
                 state.BlendHint |= CameraState.BlendHintValue.NoTransform | CameraState.BlendHintValue.NoLens;
                 PushStateToUnityCamera(ref state);
@@ -884,10 +940,11 @@ namespace Cinemachine
         private void PushStateToUnityCamera(ref CameraState state)
         {
             CurrentCameraState = state;
+            var target = ControlledObject.transform;
             if ((state.BlendHint & CameraState.BlendHintValue.NoPosition) == 0)
-                transform.position = state.FinalPosition;
+                target.position = state.FinalPosition;
             if ((state.BlendHint & CameraState.BlendHintValue.NoOrientation) == 0)
-                transform.rotation = state.FinalOrientation;
+                target.rotation = state.FinalOrientation;
             if ((state.BlendHint & CameraState.BlendHintValue.NoLens) == 0)
             {
                 Camera cam = OutputCamera;
@@ -898,7 +955,8 @@ namespace Cinemachine
                     cam.orthographicSize = state.Lens.OrthographicSize;
                     cam.fieldOfView = state.Lens.FieldOfView;
                     cam.lensShift = state.Lens.LensShift;
-                    cam.orthographic = state.Lens.Orthographic;
+                    if (state.Lens.ModeOverride != LensSettings.OverrideModes.None)
+                        cam.orthographic = state.Lens.Orthographic;
                     bool isPhysical = state.Lens.ModeOverride == LensSettings.OverrideModes.None 
                         ? cam.usePhysicalProperties : state.Lens.IsPhysicalCamera;
                     cam.usePhysicalProperties = isPhysical;
@@ -907,11 +965,7 @@ namespace Cinemachine
                         cam.sensorSize = state.Lens.SensorSize;
                         cam.gateFit = state.Lens.GateFit;
 #if CINEMACHINE_HDRP
-    #if UNITY_2019_2_OR_NEWER
                         cam.TryGetComponent<HDAdditionalCameraData>(out var hda);
-    #else
-                        var hda = cam.GetComponent<HDAdditionalCameraData>();
-    #endif
                         if (hda != null)
                         {
                             hda.physicalParameters.iso = state.Lens.Iso;