+++ Runtime/Behaviours/CinemachineFreeLook.cs
@@ -486,7 +486,7 @@ namespace Cinemachine
 
         /// Serialized in order to support copy/paste
         [SerializeField][HideInInspector][NoSaveDuringPlay]
-        private CinemachineVirtualCamera[] m_Rigs = new CinemachineVirtualCamera[3];
+        CinemachineVirtualCamera[] m_Rigs = new CinemachineVirtualCamera[3];
 
         void InvalidateRigCache() { mOrbitals = null; }
         CinemachineOrbitalTransposer[] mOrbitals = null;
@@ -667,6 +667,7 @@ namespace Cinemachine
                             if (mOrbitals[i] != null)
                             {
                                 mOrbitals[i].m_HeadingIsSlave = true;
+                                mOrbitals[i].HideOffsetInInspector = true;
                                 mOrbitals[i].m_XAxis.m_InputAxisName = string.Empty;
                                 mOrbitals[i].HeadingUpdater = UpdateXAxisHeading;
                                 mOrbitals[i].m_RecenterToTargetHeading.m_enabled = false;