+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [Unreleased]
+- Bugfix: Negative Near Clip Plane value is kept when camera is orthographic.
+- Regression fix: could not change the projection of the main camera if a CM virtual camera is active.
+- Regression fix: Axis input was ignoring CM's IgnoreTimeScale setting.
+- New feature: CinemachineBrain may control other GameObject instead of the one it is attached to.
+- Bugfix: Cinemachine assigns a default input controller delegate that returns 0 when the legacy input system is disabled.
+- Cinemachine example scenes show informative text when used with Input System instead of throwing error messages.
+
+## [2.9.0-pre.1] - 2021-10-26
 - Added ability to directly set the active blend in CinemachineBrain.
 - Bugfix: OnTargetObjectWarped() did not work properly for 3rdPersonFollow.
 - Bugfix: POV did not properly handle overridden up.
@@ -14,6 +22,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Blends were sometimes incorrect when src or dst camera is looking along world up axis.
 - Bugfix: Improve accuracy of Group Framing.
 - GameObjects created with Gameobject menu items now follow Unity naming conventions.
+- New feature: Added scene view overlay tools for Cinemachine components.
+- Regression fix: Lookahead works again.
+- Cinemachine3rdPersonAim exposes AimTarget, which is the position of where the player would hit.
 
 
 ## [2.8.0] - 2021-07-13